(chore) Split up the DatePicker into multiple components
