+++ packages/framework/esm-framework/docs/API.md
@@ -539,18 +539,6 @@ ___
 
 ___
 
-### DateInputValue
-
-Ƭ **DateInputValue**: `CalendarDate` \| `CalendarDateTime` \| `ZonedDateTime` \| `DayjsConfigType`
-
-A type for any of the acceptable date formats
-
-#### Defined in
-
-[packages/framework/esm-styleguide/src/datepicker/index.tsx:94](https://github.com/openmrs/openmrs-esm-core/blob/main/packages/framework/esm-styleguide/src/datepicker/index.tsx#L94)
-
-___
-
 ### DefaultEvaluateReturnType
 
 Ƭ **DefaultEvaluateReturnType**: `string` \| `number` \| `boolean` \| `Date` \| ``null`` \| `undefined`
@@ -1907,7 +1895,7 @@ A date picker component to select a single date. Based on React Aria, but styled
 
 #### Defined in
 
-[packages/framework/esm-styleguide/src/datepicker/index.tsx:603](https://github.com/openmrs/openmrs-esm-core/blob/main/packages/framework/esm-styleguide/src/datepicker/index.tsx#L603)
+[packages/framework/esm-styleguide/src/datepicker/OpenmrsDatePicker.tsx:94](https://github.com/openmrs/openmrs-esm-core/blob/main/packages/framework/esm-styleguide/src/datepicker/OpenmrsDatePicker.tsx#L94)
 
 ___
