+++ packages/framework/esm-styleguide/src/datepicker/datepicker.module.scss
@@ -66,8 +66,12 @@
     display: inline-flex;
   }
 
-  &:hover:not(:has([aria-disabled='true'])) {
+  &:hover:not(:disabled) {
     background-color: theme.$layer-hover;
+  }
+
+  &:disabled > svg {
+    fill: theme.$text-disabled;
   }
 
   & > svg {
@@ -114,6 +118,12 @@
     & {
       color: theme.$support-error;
     }
+  }
+
+  &[data-disabled] {
+    border-block-end: 1px solid rgba(0, 0, 0, 0);
+    color: theme.$text-disabled;
+    cursor: not-allowed;
   }
 }
