+++ packages/framework/esm-styleguide/src/datepicker/locale-context.ts
@@ -1,0 +1,5 @@
+import { createContext, useContext } from 'react';
+
+export const OpenmrsIntlLocaleContext = createContext<Intl.Locale | null>(null);
+
+export const useIntlLocale = () => useContext(OpenmrsIntlLocaleContext)!;
